=Working state. Ported to gcc and Implemented hzeller's rgb API.

+++ FontGenerator.cpp
@@ -2,6 +2,9 @@
 
 FontGenerator::FontGenerator()
 {
+
+	//sets up the vectors for hour, minutes and seconds
+	//and allocates space for PixelMatrix vector elements will point to.
 	update_count_ = 0;
 	hour_font_vector_ = new std::vector<PixelMatrix*>(HOUR_VECTOR_COUNT);
 	for (int i = 0; i < HOUR_VECTOR_COUNT; i++)
@@ -20,15 +23,15 @@ FontGenerator::FontGenerator()
 
 FontGenerator::~FontGenerator()
 {
-	for (int i = 0; i < hour_font_vector_->size(); i++)
+	for (int i = 0; i < (int) hour_font_vector_->size(); i++)
 		delete hour_font_vector_->at(i);
 	delete hour_font_vector_;
 
-	for (int i = 0; i < minute_font_vector_->size(); i++)
+	for (int i = 0; i < (int) minute_font_vector_->size(); i++)
 		delete minute_font_vector_->at(i);
 	delete minute_font_vector_;
 
-	for (int i = 0; i < second_font_vector_->size(); i++)
+	for (int i = 0; i < (int) second_font_vector_->size(); i++)
 		delete second_font_vector_->at(i);
 	delete second_font_vector_;
 
@@ -36,34 +39,39 @@ FontGenerator::~FontGenerator()
 	return;
 }
 
+//updates the vectors to hold infomation on current time
 void FontGenerator::Update()
 {
 	ClearVectors();
 
+	//getting the time using CTime
 	time_t time;
 	std::time(&time);
 	struct tm* time_info = std::localtime(&time);
 
+	//Strings hold the characters we would like in our vectors
 	std::string* hour_binary = DecToBin(time_info->tm_hour % 12);
 	std::string* minute_binary = DecToBin(time_info->tm_min);
 	std::string* second_binary = DecToBin(time_info->tm_sec);
 
-	for (int i = 0; i < hour_binary->size(); i++)
+	//pushes the character specific matricies into the appropriate vector
+	for (int i = 0; i < (int) hour_binary->size(); i++)
 	{
 		hour_font_vector_->push_back(new PixelMatrix(FONT_HEIGHT, FONT_WIDTH));
 		SetMatrixToCharacter(hour_font_vector_->at(i), hour_binary->at(i));
 	}
-	for (int i = 0; i < minute_binary->size(); i++)
+	for (int i = 0; i < (int) minute_binary->size(); i++)
 	{
 		minute_font_vector_->push_back(new PixelMatrix(FONT_HEIGHT, FONT_WIDTH));
 		SetMatrixToCharacter(minute_font_vector_->at(i), minute_binary->at(i));
 	}
-	for (int i = 0; i < second_binary->size(); i++)
+	for (int i = 0; i < (int) second_binary->size(); i++)
 	{
 		second_font_vector_->push_back(new PixelMatrix(FONT_HEIGHT, FONT_WIDTH));
 		SetMatrixToCharacter(second_font_vector_->at(i), second_binary->at(i));
 	}
 
+	//clean up
 	delete hour_binary;
 	delete minute_binary;
 	delete second_binary;
@@ -75,15 +83,16 @@ void FontGenerator::Update()
 
 void FontGenerator::ClearVectors()
 {
-	for (int i = 0; i < hour_font_vector_->size(); i++)
+	//free and zero all pointers in vectors
+	for (int i = 0; i < (int) hour_font_vector_->size(); i++)
 		delete hour_font_vector_->at(i);
 	hour_font_vector_->clear();
 
-	for (int i = 0; i < minute_font_vector_->size(); i++)
+	for (int i = 0; i < (int) minute_font_vector_->size(); i++)
 		delete minute_font_vector_->at(i);
 	minute_font_vector_->clear();
 
-	for (int i = 0; i < second_font_vector_->size(); i++)
+	for (int i = 0; i < (int) second_font_vector_->size(); i++)
 		delete second_font_vector_->at(i);
 	second_font_vector_->clear();
 
@@ -126,6 +135,10 @@ std::string* FontGenerator::DecToBin(int dec)
 {
 	std::string* bin = new std::string();
 
+	//this function reads each bit stored for the decimal number.
+	//each bit is binary OR'd with 1 for reading.
+
+	//if the decimal number is 0, push zero into string.
 	if (dec == 0)
 	{
 		bin->push_back('0');
@@ -134,11 +147,14 @@ std::string* FontGenerator::DecToBin(int dec)
 
 	while (dec > 0)
 	{
+		//if binary OR of dec against 1 equals 1, insert
+		//1 into beginning of string
 		if (dec & 1)
 			bin->insert(0, "1");
 		else
 			bin->insert(0, "0");
 
+		//shift the bits over one to read next bit
 		dec = dec >> 1;
 	}
